active_agent/requirements.txt
psutil
google-generativeai
arxiv
